Update HowToUpdateEOPLicenseFile.md
Add the following to step #3 on the filesystem method to confirm that processes stopped cleanly:
On Linux, verify that all TeamServer processes have stopped by executing ```ps aux | grep contrast``` and confirm there are no processes listed.  If myslqld is still running it may take a few minutes to terminate on its own after stopping the service.  Contact Support if it doesn't - do not kill the processes.