:sparkles: CONTRAST-32217 Automatic updates with curl example
We publish the Contrast Java agent to Maven Central. Hosting the agent
on Maven Central enables users to pull down the latest agent using a
plethora of dev and ops tools. We need to instruct users on how to
configure their system to automatically download the latest version of
the Contrast agent from Maven Central.

We will provide an example which instructs users how to configure their
Ubuntu 18.04 LTS system (with cron installed) of a shell script and the
corresponding cron configuration to demonstrate updating the Contrast
agent jar on a system on some schedule (e.g. nightly). We want to show
how to download the latest jar using a scheduled job.