Update Tomcat.md
The quotes in the environment variable SET command for Windows were misplaced - if they are placed as they were they become a literal part of the variable which leads to issues if CATALINA_OPTS is not already set to a valid value.  Using setenv.bat is by far the better option for Windows.